--- presentation/src/main/java/com/sms/presentation/main/ui/teacher_registration/component/state/Class.kt
@@ -1,7 +1,0 @@
-package com.sms.presentation.main.ui.teacher_registration.component.state
-
-enum class Class (val text: String){
-    CLASS_1("1반"),
-    CLASS_2("2반"),
-    CLASS_3("3반")
-}
