+++ presentation/src/main/java/com/sms/presentation/main/ui/teacher_registration/state/Class.kt
@@ -1,0 +1,8 @@
+package com.sms.presentation.main.ui.teacher_registration.state
+
+enum class Class (val text: String){
+    CLASS_1("1반"),
+    CLASS_2("2반"),
+    CLASS_3("3반"),
+    CLASS_4("4반")
+}
